pyproject.toml
[project]
name = "tiprankscrawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "datasets>=3.5.0",
    "lxml>=5.3.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
